backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
aiosmtplib==3.0.2
aiofiles==24.1.0
redis==5.1.1
celery[redis]==5.4.0
alembic==1.13.3
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
Jinja2==3.1.4
email-validator==2.2.0
aiosqlite==0.20.0
